requirements.txt
gradio
pytesseract
pymupdf
Pillow
transformers
torch
